more non terminals

+++ StatPyConvertor/src/main/java/gt/edu/usac/ingenieria/analyzer/errors/ErrorStpL.java
@@ -1,11 +1,11 @@
 package gt.edu.usac.ingenieria.analyzer.errors;
 
 public class ErrorStpL {
-    int line;
-    int column;
+    long line;
+    long column;
     String character;
 
-    public ErrorStpL(int line, int column, String character){
+    public ErrorStpL(long line, long column, String character){
         this.line = line;
         this.column = column;
         this.character = character;
